# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "Laravel API documentation with Swagger",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Laravel API Server"
        }
    ],
    "paths": {
...
        "/api/storePaypalInfo": {
            "post": {
                "tags": [
                    "PayPal"
                ],
                "summary": "Configure PayPal gateway",
                "description": "Store PayPal gateway configuration settings",
                "operationId": "6eea4573f0424f8e38777cbf87a4431e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "client_secret",
                                    "mode"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "string",
                                        "example": "YOUR_PAYPAL_CLIENT_ID"
                                    },
                                    "client_secret": {
                                        "type": "string",
                                        "example": "YOUR_PAYPAL_CLIENT_SECRET"
                                    },
                                    "mode": {
                                        "type": "string",
                                        "enum": [
                                            "sandbox",
                                            "live"
                                        ],
                                        "example": "sandbox"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "PayPal configuration stored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "client_id": {
                                                    "type": "string"
                                                },
                                                "mode": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "boolean"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "PayPal configuration stored successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/paypalInfo": {
            "get": {
                "tags": [
                    "PayPal"
                ],
                "summary": "Get PayPal gateway information",
                "description": "Returns the current PayPal gateway configuration settings",
                "operationId": "8d982d911c139e1ffcecce9cbaf34240",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "client_id": {
                                                    "type": "string"
                                                },
                                                "mode": {
                                                    "type": "string",
                                                    "enum": [
                                                        "sandbox",
                                                        "live"
                                                    ]
                                                },
                                                "status": {
                                                    "type": "boolean"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "PayPal configuration retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "PayPal configuration not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/updatePaypalInfo/{id}": {
            "post": {
                "tags": [
                    "PayPal"
                ],
                "summary": "Update PayPal gateway configuration",
                "description": "Updates the existing PayPal gateway configuration settings",
                "operationId": "b23d9b4af3b03148f01ea92fb7f0659c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of PayPal configuration to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client_id": {
                                        "type": "string",
                                        "example": "UPDATED_PAYPAL_CLIENT_ID"
                                    },
                                    "client_secret": {
                                        "type": "string",
                                        "example": "UPDATED_PAYPAL_CLIENT_SECRET"
                                    },
                                    "mode": {
                                        "type": "string",
                                        "enum": [
                                            "sandbox",
                                            "live"
                                        ],
                                        "example": "live"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PayPal configuration updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "client_id": {
                                                    "type": "string"
                                                },
                                                "mode": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "boolean"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "PayPal configuration updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "PayPal configuration not found"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "apiAuth": []
        }
    ]
}